+++ src/main/java/roomescape/argumentresolver/ReservationFilterArgumentResolver.java
@@ -1,4 +1,4 @@
-package roomescape.controller.argumentresolver;
+package roomescape.argumentresolver;
 
 import jakarta.servlet.http.HttpServletRequest;
 import java.time.LocalDate;
